Do not depend on y4m in the rav1e library
